Consider message edition for new Agents case
Note: We don't have yet a good way to interact with Agents when editing
an inline message. It would send the message to the current model or
agent, instead of re-sending to the original entity at the moment of
creating the message. We need to define the proper UX here.